csv-backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
pandas>=2.2,<3.0
python-decouple>=3.8,<4.0
requests>=2.32,<3.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
openpyxl>=3.1,<4.0
